backend: move tag files into tag package
