+++ frontend/src/pages/tags/TagImageListPage.tsx
@@ -11,11 +11,11 @@ import {
 import { CardActions } from "@mui/material";
 import { FC, useEffect, useState } from "react";
 import { useNavigate, useParams, useSearchParams } from "react-router";
+import { ImageFile } from "../../../bindings/github.com/michael-freling/anime-image-viewer/internal/image";
 import {
-  ImageFile,
   Tag,
-  TagService,
-} from "../../../bindings/github.com/michael-freling/anime-image-viewer/internal/image";
+  TagFrontendService,
+} from "../../../bindings/github.com/michael-freling/anime-image-viewer/internal/tag";
 import LazyImage from "../../components/LazyImage";
 
 export interface UserImages {
@@ -40,7 +40,9 @@ const TagImageListPage: FC = () => {
   });
 
   const readTag = async (tagId: string) => {
-    const response = await TagService.ReadImageFiles(parseInt(tagId, 10));
+    const response = await TagFrontendService.ReadImageFiles(
+      parseInt(tagId, 10)
+    );
     let userImages: {
       [key: number]: Array<ImageFile & { selected: boolean }>;
     } = {};
