+++ internal/tag/tags_test.go
@@ -1,0 +1,34 @@
+package tag
+
+type tagBuilder struct {
+	tags map[uint]*Tag
+}
+
+func newTagBuilder() *tagBuilder {
+	return &tagBuilder{
+		tags: make(map[uint]*Tag),
+	}
+}
+
+func (b *tagBuilder) add(tag Tag, parentID uint) *tagBuilder {
+	if parentID != 0 {
+		parent := b.tags[parentID]
+		tag.parent = parent
+		tag.FullName = parent.FullName + " > " + tag.Name
+
+		if parent.Children == nil {
+			parent.Children = []*Tag{}
+		}
+		parent.Children = append(parent.Children, &tag)
+	}
+	if tag.FullName == "" {
+		tag.FullName = tag.Name
+	}
+
+	b.tags[tag.ID] = &tag
+	return b
+}
+
+func (b tagBuilder) build(id uint) Tag {
+	return *b.tags[id]
+}
